changing the way of the less compiler integration
- added a parameter to the template.xml so that you switch it on and off, default to on
- always include a template.css and added the file
- move the logic to run the less complied in a file and made some changes to let it run

this change might not be perfect but it removes one security related problem and one performace problem in production
